feat: make recording url state and sharing way better

+++ frigate/test/test_http.py
@@ -127,10 +127,7 @@ class TestHttp(unittest.TestCase):
 
         with TestClient(app) as client:
             _insert_mock_event(id)
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            event = client.get(f"/events/{id}").json()
 
         assert event
         assert event["id"] == id
@@ -171,20 +168,11 @@ class TestHttp(unittest.TestCase):
 
         with TestClient(app) as client:
             _insert_mock_event(id)
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            event = client.get(f"/events/{id}").json()
             assert event
             assert event["id"] == id
-            client.delete(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            )
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            client.delete(f"/events/{id}", headers={"remote-role": "admin"})
+            event = client.get(f"/events/{id}").json()
             assert event == "Event not found"
 
     def test_event_retention(self):
@@ -202,25 +190,13 @@ class TestHttp(unittest.TestCase):
 
         with TestClient(app) as client:
             _insert_mock_event(id)
-            client.post(
-                f"/events/{id}/retain",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            )
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            client.post(f"/events/{id}/retain", headers={"remote-role": "admin"})
+            event = client.get(f"/events/{id}").json()
             assert event
             assert event["id"] == id
             assert event["retain_indefinitely"] is True
-            client.delete(
-                f"/events/{id}/retain",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            )
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            client.delete(f"/events/{id}/retain", headers={"remote-role": "admin"})
+            event = client.get(f"/events/{id}").json()
             assert event
             assert event["id"] == id
             assert event["retain_indefinitely"] is False
@@ -245,16 +221,13 @@ class TestHttp(unittest.TestCase):
             _insert_mock_event(morning_id, morning)
             _insert_mock_event(evening_id, evening)
             # both events come back
-            events = client.get(
-                "/events", headers={"remote-user": "admin", "remote-role": "admin"}
-            ).json()
+            events = client.get("/events").json()
             assert events
             assert len(events) == 2
             # morning event is excluded
             events = client.get(
                 "/events",
                 params={"time_range": "07:00,24:00"},
-                headers={"remote-user": "admin", "remote-role": "admin"},
             ).json()
             assert events
             # assert len(events) == 1
@@ -262,7 +235,6 @@ class TestHttp(unittest.TestCase):
             events = client.get(
                 "/events",
                 params={"time_range": "00:00,18:00"},
-                headers={"remote-user": "admin", "remote-role": "admin"},
             ).json()
             assert events
             assert len(events) == 1
@@ -294,26 +266,20 @@ class TestHttp(unittest.TestCase):
             new_sub_label_response = client.post(
                 f"/events/{id}/sub_label",
                 json={"subLabel": sub_label},
-                headers={"remote-user": "admin", "remote-role": "admin"},
+                headers={"remote-role": "admin"},
             )
             assert new_sub_label_response.status_code == 200
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            event = client.get(f"/events/{id}").json()
             assert event
             assert event["id"] == id
             assert event["sub_label"] == sub_label
             empty_sub_label_response = client.post(
                 f"/events/{id}/sub_label",
                 json={"subLabel": ""},
-                headers={"remote-user": "admin", "remote-role": "admin"},
+                headers={"remote-role": "admin"},
             )
             assert empty_sub_label_response.status_code == 200
-            event = client.get(
-                f"/events/{id}",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            ).json()
+            event = client.get(f"/events/{id}").json()
             assert event
             assert event["id"] == id
             assert event["sub_label"] == None
@@ -345,11 +311,9 @@ class TestHttp(unittest.TestCase):
             client.post(
                 f"/events/{id}/sub_label",
                 json={"subLabel": sub_label},
-                headers={"remote-user": "admin", "remote-role": "admin"},
+                headers={"remote-role": "admin"},
             )
-            sub_labels = client.get(
-                "/sub_labels", headers={"remote-user": "admin", "remote-role": "admin"}
-            ).json()
+            sub_labels = client.get("/sub_labels").json()
             assert sub_labels
             assert sub_labels == [sub_label]
 
@@ -366,9 +330,7 @@ class TestHttp(unittest.TestCase):
         )
 
         with TestClient(app) as client:
-            config = client.get(
-                "/config", headers={"remote-user": "admin", "remote-role": "admin"}
-            ).json()
+            config = client.get("/config").json()
             assert config
             assert config["cameras"]["front_door"]
 
@@ -387,10 +349,7 @@ class TestHttp(unittest.TestCase):
 
         with TestClient(app) as client:
             _insert_mock_recording(id)
-            response = client.get(
-                "/front_door/recordings",
-                headers={"remote-user": "admin", "remote-role": "admin"},
-            )
+            response = client.get("/front_door/recordings")
             assert response.status_code == 200
             recording = response.json()
             assert recording
